feat(sight): extract audit service crate
Move audit models, SQLite persistence, and risk correlation behind an
independent AuditService while keeping database opening, event transport,
HTTP, and enforcement adapters in the AgentSight host.

This preserves the existing deployment and API contracts without introducing
another daemon. Containment execution remains host-owned because it depends
on live enforcement state.
