+++ src/agentsight/src/security.rs
@@ -1,5 +1,7 @@
 //! Local security evidence, risk-case models, and SQLite persistence.
 
+use std::path::Path;
+
 #[cfg(target_os = "linux")]
 mod containment;
 #[cfg(target_os = "linux")]
@@ -30,3 +32,12 @@ pub use store::{
     ContainmentActivationResult, ContainmentClaimResult, SecurityEventStore, SecurityStore,
     SecurityStoreError,
 };
+
+pub(crate) fn open_private_store(
+    state_dir: impl AsRef<Path>,
+) -> Result<SecurityStore, SecurityStoreError> {
+    let connection =
+        crate::private_sqlite::open_private_connection(state_dir.as_ref(), "security.db")
+            .map_err(|error| SecurityStoreError::Open(error.to_string()))?;
+    SecurityStore::from_connection(connection)
+}
